Fixed it a bit.. but not enough

+++ AdventOfCode/AdventOfCode/Day11/Day11Puzzle.cs
@@ -155,5 +155,5 @@ public class Item
         WorryLevel = startingWorryLevel;
     }
     
-    public int WorryLevel { get; set; }
+    public long WorryLevel { get; set; }
 }
